Moved Firestore Documents helpers to a new file, stubbed tests
